requirements.txt
Fabric
Flask-WTF
WTForms
coverage
flask-login
